remix-of-remix-of-remix-of-emirates-sparkle-design-main/recommendation_service/requirements.txt
# Core Web
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
requests

# Recommendation & AI
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
google-generativeai
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5

# Search & Retrieval
duckduckgo-search==5.3.0
google-search-results==2.4.2

# File Parsing
pymupdf
python-docx
openpyxl
